Develop

+++ src/components/Card/Card.tsx
@@ -6,8 +6,8 @@ import { Post } from "../../utils/types/Post";
 import heart from "../../assets/Heart.png";
 import favorite from "../../assets/Favorite.png";
 import { api } from "../../utils/api";
-
 import "./styles.css";
+import { User } from "../../utils/types/User";
 
 interface CardPost {
   post: Post;
@@ -20,8 +20,11 @@ export default function Card({ post }: CardPost) {
   const [animationClass, setAnimationClass] = useState("");
 
   const dadoLocalUser = localStorage.getItem("user") || "";
-  const user = JSON.parse(dadoLocalUser);
-
+  let user: User;
+  if(dadoLocalUser){
+    user = JSON.parse(dadoLocalUser);
+  }
+  
   useEffect(() => {
     checkIsFavorite();
   }, []);
@@ -31,27 +34,29 @@ export default function Card({ post }: CardPost) {
   };
 
   const handleImageClick = () => {
-    nav(`detalhes/${post.id}`);
+    nav(`/detalhes/${post.id}`);
   };
 
   const handleFavoriteClick = async () => {
     try {
-      if (isFavorite) {
-        await api.delete(`/favorite/${post.id}`);
-      } else {
-        await api.post("/favorite", {
-          userId: user.id,
-          postId: post.id,
-        });
-      }
+      if(user){
+        if (isFavorite) {
+          await api.delete(`/favorite/${post.id}`);
+        } else {
+          await api.post("/favorite", {
+            userId: user.id,
+            postId: post.id,
+          });
+        }
 
-      setAnimationClass("animate-heart");
+        setAnimationClass("animate-heart");
 
-      setTimeout(() => {
-        setAnimationClass("");
-      }, 1000);
+        setTimeout(() => {
+          setAnimationClass("");
+        }, 1000);
 
-      setIsFavorite(!isFavorite);
+        setIsFavorite(!isFavorite);
+    } else return
     } catch (error) {
       console.error("Erro ao favoritar/desfavoritar:", error);
     }
@@ -59,12 +64,14 @@ export default function Card({ post }: CardPost) {
 
   const checkIsFavorite = async () => {
     try {
-      const response = await api.get(`/favorite/user/${user.id}`);
-  
-      const favorites = response.data;
-      const isPostFavorite = favorites.some((fav: any) => fav.postId === post.id);
-  
-      setIsFavorite(isPostFavorite);
+      if(user){
+        const response = await api.get(`/favorite/user/${user.id}`);
+    
+        const favorites = response.data;
+        const isPostFavorite = favorites.some((fav: any) => fav.postId === post.id);
+    
+        setIsFavorite(isPostFavorite);
+      } else return
     } catch (error) {
       console.error("Erro ao verificar favorito:", error);
     }
@@ -72,7 +79,7 @@ export default function Card({ post }: CardPost) {
 
   return (
     <>
-      <div className="card-div" >
+      <div className="card-div">
         <span className="card-title">{post.title}</span>
         <div className="image-div">
           <img
